Tüm Controller a yetki düzenlemesi yaptım yetki erişimi

+++ panel/application/controllers/Email_Settings.php
@@ -27,6 +27,11 @@ class Email_Settings extends MY_Controller{
     }
     public function new_form()
     {
+        if(!isAllowViewModule($this->viewTitle, "write"))
+        {
+            redirect(base_url($this->viewTitle));
+            die();
+        }
         $viewData = new stdClass();
         $viewData->viewTitle = $this->viewTitle;
         $viewData->viewFolder = $this->viewFolder;
@@ -35,6 +40,11 @@ class Email_Settings extends MY_Controller{
     }
     public function save()
     {
+        if(!isAllowViewModule($this->viewTitle, "write"))
+        {
+            redirect(base_url($this->viewTitle));
+            die();
+        }
         $this->load->library("form_validation");
         $this->form_validation->set_rules("protocol","Protokol Numarası", "required|trim");
         $this->form_validation->set_rules("host","E-posta Sunucusu", "required|trim");
@@ -99,6 +109,11 @@ class Email_Settings extends MY_Controller{
     }
     public function update_form($id)
     {
+        if(!isAllowViewModule($this->viewTitle, "update"))
+        {
+            redirect(base_url($this->viewTitle));
+            die();
+        }
         $item = $this->email_setting_model->get(
             array(
                 "id"    => $id
@@ -113,6 +128,11 @@ class Email_Settings extends MY_Controller{
     }
     public function update($id)
     {
+        if(!isAllowViewModule($this->viewTitle, "update"))
+        {
+            redirect(base_url($this->viewTitle));
+            die();
+        }
         $this->load->library("form_validation");
         $this->form_validation->set_rules("protocol","Protokol Numarası", "required|trim");
         $this->form_validation->set_rules("host","E-posta Sunucusu", "required|trim");
@@ -183,6 +203,11 @@ class Email_Settings extends MY_Controller{
     }
     public function delete($id)
     {   
+        if(!isAllowViewModule($this->viewTitle, "delete"))
+        {
+            redirect(base_url($this->viewTitle));
+            die();
+        }
         $delete = $this->email_setting_model->delete(
             array("id" => $id)
         );
@@ -207,6 +232,11 @@ class Email_Settings extends MY_Controller{
     }
     public function isActiveSetter($id)
     {
+        if(!isAllowViewModule($this->viewTitle, "update"))
+        {
+            redirect(base_url($this->viewTitle));
+            die();
+        }
         if($id)
 		{
 			$isActive = ($this->input->post("data") === "true") ? 1 : 0;
