+++ panel/application/controllers/Portfolio_categories.php
@@ -28,6 +28,11 @@ class Portfolio_categories extends MY_Controller{
     }
     public function new_form()
     {
+        if(!isAllowViewModule($this->viewTitle, "write"))
+        {
+            redirect(base_url($this->viewTitle));
+            die();
+        }
         $viewData = new stdClass();
         $viewData->viewTitle = $this->viewTitle;
         $viewData->viewFolder = $this->viewFolder;
@@ -36,6 +41,11 @@ class Portfolio_categories extends MY_Controller{
     }
     public function save()
     {
+        if(!isAllowViewModule($this->viewTitle, "write"))
+        {
+            redirect(base_url($this->viewTitle));
+            die();
+        }
         $this->load->library("form_validation");
         $this->form_validation->set_rules("title","Başlık","required|trim");
         $this->form_validation->set_message(
@@ -87,6 +97,11 @@ class Portfolio_categories extends MY_Controller{
     }
     public function update_form($id)
     {
+        if(!isAllowViewModule($this->viewTitle, "update"))
+        {
+            redirect(base_url($this->viewTitle));
+            die();
+        }
         $viewData = new stdClass();
         $item = $this->portfolio_category_model->get(
             array(
@@ -101,6 +116,11 @@ class Portfolio_categories extends MY_Controller{
     }
     public function update($id)
     {
+        if(!isAllowViewModule($this->viewTitle, "update"))
+        {
+            redirect(base_url($this->viewTitle));
+            die();
+        }
         $this->load->library("form_validation");
         $this->form_validation->set_rules("title", "Başlık", "required|trim");
         $this->form_validation->set_message(
@@ -155,6 +175,11 @@ class Portfolio_categories extends MY_Controller{
     }
     public function isActiveSetter($id)
     {
+        if(!isAllowViewModule($this->viewTitle, "update"))
+        {
+            redirect(base_url($this->viewTitle));
+            die();
+        }
         if($id)
         {
             $isActive = ($this->input->post("data") === "true") ? 1 : 0;
@@ -170,6 +195,11 @@ class Portfolio_categories extends MY_Controller{
     }
     public function delete($id)
     {
+        if(!isAllowViewModule($this->viewTitle, "delete"))
+        {
+            redirect(base_url($this->viewTitle));
+            die();
+        }
         $delete = $this->portfolio_category_model->delete(
             array("id" => $id)
         );
