+++ panel/application/controllers/Portfolios.php
@@ -33,6 +33,11 @@ class Portfolios extends MY_Controller {
 	}
 	public function new_form()
 	{
+		if(!isAllowViewModule($this->viewTitle, "write"))
+        {
+            redirect(base_url($this->viewTitle));
+            die();
+        }
 		$viewData = new stdClass();
 		$viewData->viewTitle = $this->viewTitle;
 		$viewData->viewFolder = $this->viewFolder;
@@ -43,6 +48,11 @@ class Portfolios extends MY_Controller {
 	}
 	public function save()
 	{
+		if(!isAllowViewModule($this->viewTitle, "write"))
+        {
+            redirect(base_url($this->viewTitle));
+            die();
+        }
 		$this->load->library("form_validation");
 		//kuralların yazıldığı alan
 		$this->form_validation->set_rules("title","Başlık","required|trim");
@@ -101,6 +111,11 @@ class Portfolios extends MY_Controller {
 	}
 	public function update_form($id)
 	{
+		if(!isAllowViewModule($this->viewTitle, "update"))
+        {
+            redirect(base_url($this->viewTitle));
+            die();
+        }
 		$categories = $this->portfolio_category_model->get_all(array("isActive" => 1));
 		$item = $this->portfolio_model->get(
 			array(
@@ -117,6 +132,11 @@ class Portfolios extends MY_Controller {
 	}
 	public function update($id)
 	{
+		if(!isAllowViewModule($this->viewTitle, "update"))
+        {
+            redirect(base_url($this->viewTitle));
+            die();
+        }
 		$this->load->library("form_validation");
 		//kuralların yazıldığı alan
 		$this->form_validation->set_rules("title","Başlık","required|trim");
@@ -182,6 +202,11 @@ class Portfolios extends MY_Controller {
 	}
 	public function delete($id)
 	{
+		if(!isAllowViewModule($this->viewTitle, "delete"))
+        {
+            redirect(base_url($this->viewTitle));
+            die();
+        }
 		$delete = $this->portfolio_model->delete(
 			array(
 				"id" => $id
@@ -233,6 +258,11 @@ class Portfolios extends MY_Controller {
 	}
 	public function imageDelete($id, $parent_id)
 	{
+		if(!isAllowViewModule($this->viewTitle, "delete"))
+        {
+            redirect(base_url($this->viewTitle));
+            die();
+        }
 		$fileName = $this->portfolio_image_model->get(
 			array(
 				"id" => $id
@@ -255,6 +285,11 @@ class Portfolios extends MY_Controller {
 	}
 	public function isActiveSetter($id)
 	{
+		if(!isAllowViewModule($this->viewTitle, "update"))
+        {
+            redirect(base_url($this->viewTitle));
+            die();
+        }
 		if($id)
 		{
 			$isActive = ($this->input->post("data") === "true") ? 1 : 0;
@@ -270,6 +305,11 @@ class Portfolios extends MY_Controller {
 	}
 	public function isCoverSetter($id, $parent_id)
 	{
+		if(!isAllowViewModule($this->viewTitle, "update"))
+        {
+            redirect(base_url($this->viewTitle));
+            die();
+        }
 		if($id && $parent_id)
 		{
 			$isCover = ($this->input->post("data") === "true") ? 1 : 0;
@@ -312,6 +352,11 @@ class Portfolios extends MY_Controller {
 	}
 	public function rankSetter()
 	{
+		if(!isAllowViewModule($this->viewTitle, "update"))
+        {
+            redirect(base_url($this->viewTitle));
+            die();
+        }
 		$data = $this->input->post("data");
 		parse_str($data,$order);
 		$items = $order["ord"];
@@ -330,6 +375,11 @@ class Portfolios extends MY_Controller {
 	}
 	public function rankSetterImage()
 	{
+		if(!isAllowViewModule($this->viewTitle, "update"))
+        {
+            redirect(base_url($this->viewTitle));
+            die();
+        }
 		$data = $this->input->post("data");
 		parse_str($data,$order);
 		$items = $order["ord"];
@@ -348,6 +398,11 @@ class Portfolios extends MY_Controller {
 	}
 	public function image_form($id)
 	{
+		if(!isAllowViewModule($this->viewTitle, "write"))
+        {
+            redirect(base_url($this->viewTitle));
+            die();
+        }
 		$item = $this->portfolio_model->get(
 			array(
 				"id" => $id
@@ -368,6 +423,11 @@ class Portfolios extends MY_Controller {
 	}
 	public function image_upload($id)
 	{
+		if(!isAllowViewModule($this->viewTitle, "write"))
+        {
+            redirect(base_url($this->viewTitle));
+            die();
+        }
 		$file_name = CharConvert(pathinfo($_FILES["file"]["name"], PATHINFO_FILENAME)). "." . pathinfo($_FILES["file"]["name"], PATHINFO_EXTENSION);
 		$config["allowed_types"] = "jpg|jpeg|png";
 		$config["upload_path"] = "uploads/$this->viewFolder/";
@@ -394,7 +454,8 @@ class Portfolios extends MY_Controller {
 			echo "İşlem Başarısız";
 		}
 	}
-	public function refresh_image_list($id){
+	public function refresh_image_list($id)
+	{
 
         $viewData = new stdClass();
 
@@ -416,6 +477,11 @@ class Portfolios extends MY_Controller {
     }
 	public function isActiveSetterImage($id)
 	{
+		if(!isAllowViewModule($this->viewTitle, "update"))
+        {
+            redirect(base_url($this->viewTitle));
+            die();
+        }
 		if($id)
 		{
 			$isActive = ($this->input->post("data") === "true") ? 1 : 0;
